requirements.txt
streamlit==1.41.1
pandas==2.2.3
numpy==1.26.4
openpyxl==3.1.5
xlsxwriter==3.2.5
ipython>=7.0.0,<8.0.0
setuptools==75.1.0
groq==0.29.0
dotenv
azure-storage-blob
